# contracts/admin.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PDF Master staff API",
    "version": "1.1.0",
    "description": "Separate staff-session realm with role authorization and CSRF protection. Read-only analytics APIs and server-rendered staff forms. Secrets are write-only except one-time TOTP enrollment. Do not distribute to the customer app."
  },
  "servers": [
    {
      "url": "/ops/api/v1"
    }
  ],
  "paths": {
...
    "/analytics/revenue": {
      "servers": [
        {
          "url": "/ops"
        }
      ],
      "get": {
        "operationId": "getRevenueReport",
        "summary": "Aggregate finance for Analyst; identifiable rows and refund links only for Finance/Administrator. Refunds use their own confirmation dates.",
        "x-allowed-roles": [
          "Analyst",
          "Finance",
          "Administrator"
        ],
        "security": [
          {
            "staffSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Server-rendered HTML, Cache-Control: no-store."
          },
          "302": {
            "description": "Staff authentication required."
          },
          "403": {
            "description": "Role denied."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ops_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRFToken",
        "description": "Django CSRF token, obtained from the staff form csrfmiddlewaretoken or CSRF cookie. Native browser forms may send the hidden csrfmiddlewaretoken form field instead."
      }
    }
  }
}